feat: 좌표 간 거리 계산하는 DistanceCalculator 클래스 구현
